# lib/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a REST server for ioBroker."
  version: "2.0.3"
  title: "ioBroker Swagger UI"
  contact:
    email: "admin@iobroker.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
securityDefinitions:
  basicAuth:
    type: basic
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your bearer token in the format **Bearer &lt;token&gt;**"
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /object/{objectId}:
    x-swagger-router-controller: object
    get:
      tags:
        - "object"
      summary: "Read object"
      description: ""
      operationId: "readObject"
      produces:
        - "application/json"
      parameters:
        - name: "objectId"
          in: "path"
          description: "ID of the object to return"
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Object"
        404:
          description: "Object not found"
        400:
          description: "Invalid object ID supplied"
    post:
      tags:
        - "object"
      summary: "Create the object"
      description: "Create new object. If object already exists, the error will be returned"
      operationId: "createObject"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "objectId"
          description: "ID of the object to write"
          required: true
          type: "string"
        - in: "body"
          name: "value"
          description: "Object"
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Object"
        400:
          description: "Invalid object ID supplied"
        409:
          description: "Object already exists"
        405:
          description: "Invalid object structure"
    put:
      tags:
        - "object"
      summary: "Update the object"
      description: "To delete the parts of the object, set the value to 'null', like {\"common\": {\"desc\": null} to delete the \"common.desc\". If object does not exist it will be created"
      operationId: "updateObject"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "objectId"
          description: "ID of the object to write"
          required: true
          type: "string"
        - in: "body"
          name: "value"
          description: "Object"
          required: true
          schema:
            $ref: "#/definitions/Object"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Object"
        400:
          description: "Invalid object ID supplied"
        404:
          description: "Object not found"
        405:
          description: "Invalid object structure"
    delete:
      tags:
        - "object"
      summary: "Delete object"
      description: ""
      operationId: "deleteObject"
      produces:
        - "application/json"
      parameters:
        - name: "objectId"
          in: "path"
          description: "ID of the object to be deleted"
          required: true
          type: "string"
      responses:
        200:
          description: "successful operation"
        404:
          description: "Object not found"
        400:
          description: "Invalid object ID supplied"
  /object/{objectId}/subscribe:
    x-swagger-router-controller: object
    post:
      tags:
        - "object"
      summary: "Subscribe on object changes"
      description: "Subscribe on object changes"
      operationId: "subscribeObject"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "objectId"
          in: "path"
          description: "ID of the state to subscribe"
          required: true
          type: "string"
        - name: "url"
          in: "body"
          description: "URL of web hook"
          required: true
          schema:
            $ref: "#/definitions/UrlHook"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Object"
        404:
          description: "Object not found"
        400:
          description: "Invalid object ID supplied"
    delete:
      tags:
        - "object"
      summary: "Unsubscribe from object changes"
      description: "Unsubscribe on object changes"
      operationId: "unsubscribeObject"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "objectId"
          in: "path"
          description: "ID of the object to unsubscribe"
          required: true
          type: "string"
        - name: "subscribe"
          in: "body"
          description: "Subscribe request"
          required: true
          schema:
            $ref: "#/definitions/UrlHook"
      responses:
        200:
          description: "successful operation"
        404:
          description: "Object not found"
        400:
          description: "Invalid object ID supplied"
  /objects:
    x-swagger-router-controller: object
    get:
      tags:
        - "object"
      summary: "Get list of objects"
      operationId: "listObjects"
      produces:
        - "application/json"
      parameters:
        - name: "filter"
          in: "query"
          description: "Filter, like modbus.0.*"
          required: true
          type: "string"
        - name: "type"
          in: "query"
          description: "Type of objects"
          required: false
          type: "string"
          enum: [state, channel, device, enum, instance, adapter, host, chart, script, folder]
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/ObjectArray"
        404:
          description: "URL or session not found"
        422:
          description: "Invalid parameters supplied"
  /objects/subscribe:
    x-swagger-router-controller: object
    get:
      tags:
        - "state"
      summary: "Read current subscribes"
      operationId: "getObjectsSubscribes"
      produces:
        - "application/json"
      parameters:
        - name: "method"
          in: "query"
          required: true
          description: "method=polling forces subscribe via long polling"
          enum: [ polling, POST, GET, PUT, PATCH ]
          type: "string"
        - name: "url"
          description: "The same URL as by subscription. By polling it must be omitted"
          in: "query"
          type: "string"
        - name: "sid"
          in: "query"
          type: "string"
          description: "Only if method=polling: session ID"
      responses:
        200:
          description: "successful operation"
        404:
          description: "URL or session not found"
        422:
          description: "Invalid parameters supplied"
    post:
      tags:
        - "object"
      summary: "Subscribe for objects updates"
      operationId: "subscribeObjects"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "Subscribe"
          in: "body"
          description: "URL of web hook and the subscribe pattern"
          required: true
          schema:
            $ref: "#/definitions/UrlHookWithPattern"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid filter supplied"
    delete:
      tags:
        - "object"
      summary: "Unsubscribe from object updates"
      description: "If no filter provided, all the object subscribes will be cleared for this web hook"
      operationId: "unsubscribeObjects"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "Unsubscribe"
          in: "body"
          description: "URL of web hook and the unsubscribe pattern"
          required: true
          schema:
            $ref: "#/definitions/UrlHook"
      responses:
        200:
          description: "successful operation"
        400:
          description: "Invalid filter supplied"
...
definitions:
...
  Object:
    type: "object"
    properties:
      _id:
        type: "string"
        description: "ID of the state"
      type:
        type: "string"
        description: "Type of this object"
        enum: [state, channel, device, enum, instance, adapter, host, chart, script, folder]
      common:
        type: "object"
        description: "Common state description"
      native:
        type: "object"
        description: "Native state description"
  ObjectArray:
    type: "object"
    additionalProperties:
      $ref: "#/definitions/Object"
  UrlHook:
    type: "object"
    properties:
      url:
        type: "string"
        description: "URL"
      method:
        type: "string"
        description: "HTTP Method"
        enum: [ polling, POST, GET, PUT, PATCH ]
  UrlHookWithPattern:
    type: "object"
    required: [pattern]
    properties:
      url:
        type: "string"
        description: "URL"
      onchange:
        type: "boolean"
        description: "If the updates must be delivered if the value changed"
      pattern:
        type: "string"
        description: "Subscribe pattern like: \"system.adapter.admin.0.*\""
      method:
        type: "string"
        description: "HTTP Method"
        enum: [ polling, POST, GET, PUT, PATCH ]
...